Create util package in main_screen package and move Format.kt file to it
