--- app/src/main/java/com/sadegh/calculator/homeScreen/.idea/.gitignore
@@ -1,3 +1,0 @@
-# Default ignored files
-/shelf/
-/workspace.xml
